Reconfigure (proofs) preferences structure
so that "Proofs" tab at "Explanation" Protege preferences section is a
section as well now. All the settings are "General" tab in that section.
There is a list of proof explanation plugins at the tab also. Any proof
plugin is able to use corresponding proof explanation preferences
extension point in case it has its own preferences tab to add.